Set up killProcess to ensure PhantomJS is stopped, including children
It seems that `proc_terminate` doesn't reliably shut down child
processes, especially in Windows.

A new method of `killProcess` has been set up that will use Windows
specific `taskkill` and Unix's `kill` to stop the process and its
children.

This in the aim to resolve issue #20 and to mitigate issue #25.