Allow Phantoman to only start for specific suites
This adds the `suites` configuration ability which allows one to set
what suites to start PhantomJS for.

If *any* of the current suites being run match any of the desired
suites, PhantomJS is started for the *entire* test run and is stopped
after the last suite.

Suite configuration can either be a single string or an array of values.

Many thanks to @acuthbert for initially getting this started.

(Fixes #3)